third_party/blink/web_tests/media/W3C/audio/events/event_timeupdate-expected.txt

setting src attribute on a sufficiently long autoplay video should trigger timeupdate event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
